refactor(frontend): consolidate context detail into tabbed layout (#232)
Replace 4 separate context detail routes (stats, graph, settings,
search-settings) with a single tabbed page at /contexts/[id] using
Entity Tabs (pill style) and useTabParam for URL sync.

- Extract OverviewTabPanel, ConnectionsTabPanel, SettingsTabPanel,
  SearchSettingsSection as self-contained tab panel components
- Merge Settings + Search Settings into one Settings tab with Separator
- Convert old routes to redirect stubs (302 → new route with ?tab=)
- Update all internal links (contexts list, dashboard, ContextSelector)
- Add contextDetail i18n namespace (en + ja)
