+++ frontend/src/app/(authenticated)/workspace/contexts/page.tsx
@@ -203,7 +203,7 @@ export default function ContextsPage() {
     const editId = searchParams.get("edit");
     if (editId && !editHandled.current) {
       editHandled.current = true;
-      router.replace(`/workspace/contexts/${editId}/settings`);
+      router.replace(`/workspace/contexts/${editId}?tab=settings`);
     }
   }, [searchParams, router]);
 
@@ -357,7 +357,7 @@ export default function ContextsPage() {
   };
 
   const handleViewStats = async (context: Context) => {
-    router.push(`/workspace/contexts/${context.id}/stats`);
+    router.push(`/workspace/contexts/${context.id}`);
   };
 
   const handleLoadStats = async (context: Context) => {
@@ -1047,7 +1047,7 @@ export default function ContextsPage() {
                               <DropdownMenuItem
                                 onClick={() =>
                                   router.push(
-                                    `/workspace/contexts/${context.id}/settings`,
+                                    `/workspace/contexts/${context.id}?tab=settings`,
                                   )
                                 }
                               >
@@ -1060,7 +1060,7 @@ export default function ContextsPage() {
                           <DropdownMenuItem
                             onClick={() =>
                               router.push(
-                                `/workspace/contexts/${context.id}/search-settings`,
+                                `/workspace/contexts/${context.id}?tab=settings`,
                               )
                             }
                           >
